Add absolutely positioned wrapper to allow using `position: sticky` on the sidebar
The previous code depended on adding the scroll position to the root element as a custom property.
This resulted in excessive style recalculation which made scrolling laggy.

The api pages, which use separate stylesheets, actually solve this in a good way, by using a grid to define the sidebar area to be next to the main area.

Making the same markup change here results in quite a bit of things needing to be updated.

This commits attempts to apply a similar fix in a non intrusive manner, allowing the page to keep on working the way it did otherwise, but already removing the big performance cost.